replacing CAN with Serial LED controller
