--- src/main/java/prime/can/Color.java
@@ -1,14 +1,0 @@
-package prime.can;
-
-public class Color {
-
-  public byte r;
-  public byte g;
-  public byte b;
-
-  public Color(byte r, byte g, byte b) {
-    this.r = r;
-    this.g = g;
-    this.b = b;
-  }
-}
